added Dijkstra algorithm

+++ lib/result_struct.h
@@ -6,7 +6,7 @@
 struct result{
     uint16_t start;             // start vertex of edge
     uint16_t end;               // end vertex of edge
-    int16_t weight;             // weight of edge
+    int32_t weight;             // weight of edge
     struct result* next;        // pointer to next edge
 };
